trim comments, and respect trim applied in user-queries.
grammars like markdown extend the range of many nodes to the next line,
which may be out of bounds, naively. In order to circumvent this/make
handling it easier, we trim where we are in control of the query, and
respect trim applied by the user where we are not.