fix: manually handle split-up `TSTree:included_ranges`.
in markdown-codeblocks, TSTree:included_ranges() does not return one
range for the entire tree, but has one range for every line.
While this is compatible with regular parsing,
LanguageTree:tree_for_range() does not handle this correctly, and does
not see these ranges as one big range, but many small ones.

While this should probably be fixed in neovim itself, using a fixed
implementation here is quicker and works on more (older) versions.