when merging injection-ranges, allow whitespace to connect ranges.
The markdown-parser has injected ranges that exclude the indentation of
the fenced code block.
This means that connecting ranges end-to-start does not work (the
indentation is between them).
While very exact, it kind of messes up what we're doing here. To fix it,
we treat ranges as adjacent even if there is whitespace between them.